refactor: assemble main js files into a single bundle
