+++ com/templates/com/screen_slideshow.jinja
@@ -4,7 +4,7 @@
     <title>{% trans %}Slideshow{% endtrans %}</title>
     <link rel="shortcut icon" href="{{ static('core/img/favicon.ico') }}">
     <link href="{{ static('css/slideshow.scss') }}" rel="stylesheet" type="text/css" />
-    <script type="module" src="{{ static('bundled/alpine-index.ts') }}"></script>
+    <script type="module" src="{{ static('bundled/base-bundle-index.ts') }}"></script>
     <script type="module" src="{{ static('bundled/com/slideshow-index.ts') }}"></script>
   </head>
   <body x-data="slideshow([
