--- core/static/bundled/country-flags-index.ts
@@ -1,3 +1,0 @@
-import { polyfillCountryFlagEmojis } from "country-flag-emoji-polyfill";
-
-polyfillCountryFlagEmojis();
