+++ package.json
@@ -29,6 +29,7 @@
     "@biomejs/biome": "^2.4.15",
     "@hey-api/openapi-ts": "^0.94.5",
     "@types/alpinejs": "^3.13.11",
+    "@types/alpinejs__sort": "^3.13.0",
     "@types/cytoscape-cxtmenu": "^3.4.5",
     "@types/cytoscape-klay": "^3.1.5",
     "@types/js-cookie": "^3.0.6",
